+++ src/GuiQt/AnnotationCoordinateWidget.cxx
@@ -470,9 +470,9 @@ AnnotationCoordinateWidget::valueChanged()
                 }
                 else {
                     float xyz[3] = {
-                        m_xCoordSpinBox->value(),
-                        m_yCoordSpinBox->value(),
-                        m_zCoordSpinBox->value()
+                        (float)m_xCoordSpinBox->value(),
+                        (float)m_yCoordSpinBox->value(),
+                        (float)m_zCoordSpinBox->value()
                     };
                     coordinateCopy.setXYZ(xyz);
                 }
